# [Backport 6.2] replica/table: check memtable before discarding tombstone during read