@scylladb-promoter 2024-10-23
#### :green_circle: CI State: SUCCESS

:white_check_mark: - [Build](https://jenkins.scylladb.com/job/scylla-6.2/job/scylla-ci/45/)
:white_check_mark: - [Unit Tests Custom](https://jenkins.scylladb.com/job/scylla-6.2/job/scylla-ci/45/)
**The following new/updated tests ran 100 times for each mode:**
:small_blue_diamond: topology_custom/test_compacting_reader_tombstone_gc
:white_check_mark: - [Container Test](https://jenkins.scylladb.com/job/scylla-6.2/job/scylla-ci/45/)
:white_check_mark: - [dtest with tablets](https://jenkins.scylladb.com/job/scylla-6.2/job/gating-dtest-release-with-tablets/95/)
:white_check_mark: - [dtest with topology changes](https://jenkins.scylladb.com/job/scylla-6.2/job/gating-dtest-release-with-consistent-topology-changes/97/)
:white_check_mark: - [dtest](https://jenkins.scylladb.com/job/scylla-6.2/job/gating-dtest-release/100/)
:white_check_mark: - [Unit Tests](https://jenkins.scylladb.com/job/scylla-6.2/job/scylla-ci/45/)
#### Build Details:
- Duration: 6 hr 42 min
- Builder: i-0dd78916a49be608d (m5ad.8xlarge)