@denesb 2024-10-25
Closed by https://github.com/scylladb/scylladb/commit/3f04df55ebbd4b732b55bc003cd6a9ff32889dc1.